Refactoring to provide a better multiplatform toolchain
The code was relying a lot in "is_windows" parameter everywhere, so this was changed to defined all repositories and allow toolchain resolution to pick the right one.

Toolchain will now have functionality to create a runner or run nushell using ctx.actions.run easily.

nu_binary will now be able to produce a binary file that can be run via:
1) bazel run
2) directly via bazel-bin\binary.sh
3) ctx.actions.run on a dependant rule